STRIPE/STRIPE BACK/stripe-back/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.5.0
stripe==7.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-cloud-firestore==2.14.0
Werkzeug==3.0.1
